move indicator to indicators.py
